fix issue while merge code

+++ src/main/java/com/learning/btmlearning/repository/CategoryRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     boolean existsByName(String name);
     boolean existsBySlug(String slug);
-    boolean existsBySlugAndIdNot(String slug, Integer id);
+    boolean existsBySlugAndIdNot(String slug, Long id);
 }
